LayoutTests/platform/gtk/accessibility/aria-roles-unignored-expected.txt

Simple paragraph

A paragraph pretending to be a table

A label Who said label? It's a heading!
A form with a button Click me!
Just a button Click me!
Just some text inside a div
This div is contains a textbox (an entry)
This tests that ARIA roles are not ignored for 'p','label', 'form' and 'div' elements

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS element.role is 'AXRole: paragraph'
PASS element.role is 'AXRole: table'
PASS element.role is 'AXRole: label'
PASS element.role is 'AXRole: heading'
PASS element.role is 'AXRole: form'
PASS element.role is 'AXRole: push button'
PASS element.role is 'AXRole: section'
PASS element.role is 'AXRole: entry'
PASS successfullyParsed is true

TEST COMPLETE
